requirements.txt
# Bifrost API - FastAPI Stack Requirements

# =============================================================================
# Web Framework
# =============================================================================
fastapi
uvicorn[standard]
python-multipart  # Form data handling

# =============================================================================
# Database (PostgreSQL)
# =============================================================================
sqlalchemy[asyncio]
asyncpg  # Async PostgreSQL driver
psycopg2-binary  # Sync PostgreSQL driver (for SDK/workflow threads)
alembic  # Database migrations
greenlet  # Required by SQLAlchemy async

# =============================================================================
# Message Queue (RabbitMQ)
# =============================================================================
aio-pika  # Async RabbitMQ client

# =============================================================================
# Cache & Sessions (Redis)
# =============================================================================
redis  # Includes async support (aioredis merged in v5.0)

# =============================================================================
# Authentication & Security
# =============================================================================
pyjwt[crypto]  # JWT tokens
pwdlib[bcrypt]  # Modern password hashing
cryptography  # Encryption utilities
pyotp  # TOTP/HOTP for MFA
webauthn>=2.0.0  # WebAuthn/Passkeys support

# =============================================================================
# Configuration
# =============================================================================
pydantic
pydantic-settings
python-dotenv
email-validator  # Required by Pydantic for email validation

# =============================================================================
# HTTP & Async
# =============================================================================
httpx  # Async HTTP client
aiohttp
aiofiles

# =============================================================================
# Data Processing
# =============================================================================
PyYAML
python-dateutil
croniter  # CRON expression parsing
jmespath

# =============================================================================
# Code Transformation
# =============================================================================
libcst  # Concrete Syntax Tree for Python - preserves formatting when modifying code

# =============================================================================
# Templating
# =============================================================================
jinja2

# =============================================================================
# Git Integration
# =============================================================================
dulwich
PyGithub
GitPython

# =============================================================================
# Background Jobs
# =============================================================================
apscheduler  # Scheduler

# =============================================================================
# Logging & Monitoring
# =============================================================================
python-json-logger
structlog
psutil  # Process monitoring for worker heartbeats

# =============================================================================
# Testing
# =============================================================================
pytest
pytest-asyncio
pytest-cov
pytest-env
pytest-timeout  # Per-test timeout protection (catches infinite hangs)
websockets  # For E2E WebSocket tests

# =============================================================================
# Object Storage (S3/MinIO)
# =============================================================================
aiobotocore  # Async S3 client
types-aiobotocore[s3]  # Type stubs for S3

# =============================================================================
# LLM Providers (AI Agent System)
# =============================================================================
openai  # OpenAI API client
anthropic  # Anthropic Claude API client

# =============================================================================
# MCP (Model Context Protocol)
# =============================================================================
mcp>=1.26.0  # Model Context Protocol SDK - structuredContent support requires 1.26+
fastmcp>=3.2.0,<4  # FastMCP for MCP server (internal + external access)

# =============================================================================
# Vector Store (Knowledge/RAG)
# =============================================================================
pgvector  # PostgreSQL vector extension for semantic search

# =============================================================================
# CLI / TUI
# =============================================================================
textual  # TUI framework used by bifrost CLI (file_select, progress, sync, watch)

# =============================================================================
# Development
# =============================================================================
debugpy  # Remote debugging (VS Code attach)
ruff
watchdog  # File watching for hot reload (watchmedo command)
pathspec  # gitignore-style file matching (used by bifrost CLI watch handler tests)
